*** empty log message ***

+++ rose/modules/Rose-DB-Object/t/db-object-std-cached.t
@@ -2,7 +2,7 @@
 
 use strict;
 
-use Test::More tests => 200;
+use Test::More tests => 241;
 
 BEGIN 
 {
@@ -10,7 +10,7 @@ BEGIN
   use_ok('Rose::DB::Object::Std::Cached');
 }
 
-our($PG_HAS_CHKPASS, $HAVE_PG, $HAVE_MYSQL, $HAVE_INFORMIX);
+our($PG_HAS_CHKPASS, $HAVE_PG, $HAVE_MYSQL, $HAVE_INFORMIX, $HAVE_SQLITE);
 
 #
 # Postgres
@@ -389,6 +389,112 @@ SKIP: foreach my $db_type ('informix')
   ok($@, 'alias_column() nonesuch');
 }
 
+#
+# SQLite
+#
+
+SKIP: foreach my $db_type ('sqlite')
+{
+  skip("SQLite tests", 41)  unless($HAVE_SQLITE);
+
+  Rose::DB->default_type($db_type);
+
+  my $of = MySQLiteObject->new(name => 'John');
+
+  ok(ref $of && $of->isa('MySQLiteObject'), 'cached new() 1');
+
+  ok($of->save, 'save() 1');
+
+  my $of2 = MySQLiteObject->new(id => $of->id);
+
+  ok(ref $of2 && $of2->isa('MySQLiteObject'), 'cached new() 2');
+
+  ok($of2->load, 'cached load()');
+
+  is($of2->name, $of->name, 'load() verify 1');
+
+  my $of3 = MySQLiteObject->new(id => $of2->id);
+
+  ok(ref $of3 && $of3->isa('MySQLiteObject'), 'cached new() 3');
+
+  ok($of3->load, 'cached load()');
+
+  is($of3->name, $of2->name, 'cached load() verify 2');
+
+  is($of3, $of2, 'load() verify cached 1');
+  is($of2, $of, 'load() verify cached 2');
+
+  is(keys %MySQLiteObject::Objects_By_Id, 1, 'cache check 1');
+
+  ok($of->forget, 'forget()');
+
+  is(keys %MySQLiteObject::Objects_By_Id, 0, 'cache check 2');
+
+  # Standard tests
+
+  my $o = MySQLiteObject->new(name => 'John');
+
+  ok(ref $o && $o->isa('MySQLiteObject'), "new() 1 - $db_type");
+
+  $o->flag2('true');
+  $o->date_created('now');
+  $o->last_modified($o->date_created);
+  $o->save_col(22);
+
+  ok($o->save, "save() 1 - $db_type");
+  ok($o->load, "load() 1 - $db_type");
+
+  my $o2 = MySQLiteObject->new(id => $o->id);
+
+  ok(ref $o2 && $o2->isa('MySQLiteObject'), "new() 2 - $db_type");
+
+  is($o2->bits->to_Bin, '00101', "bits() (bitfield default value) - $db_type");
+
+  ok($o2->load, "load() 2 - $db_type");
+  ok(!$o2->not_found, "not_found() 1 - $db_type");
+
+  is($o2->name, $o->name, "load() verify 1 - $db_type");
+  is($o2->date_created, $o->date_created, "load() verify 2 - $db_type");
+  is($o2->last_modified, $o->last_modified, "load() verify 3 - $db_type");
+  is($o2->status, 'active', "load() verify 4 (default value) - $db_type");
+  is($o2->flag, 1, "load() verify 5 (default boolean value) - $db_type");
+  is($o2->flag2, 1, "load() verify 6 (boolean value) - $db_type");
+  is($o2->save_col, 22, "load() verify 7 (aliased column) - $db_type");
+  is($o2->start->ymd, '1980-12-24', "load() verify 8 (date value) - $db_type");
+
+  is($o2->bits->to_Bin, '00101', "load() verify 9 (bitfield value) - $db_type");
+
+  $o2->name('John 2');
+  $o2->start('5/24/2001');
+
+  sleep(1); # keep the last modified dates from being the same
+
+  $o2->last_modified('now');
+  ok($o2->save, "save() 2 - $db_type");
+  ok($o2->load, "load() 3 - $db_type");
+
+  is($o2->date_created, $o->date_created, "save() verify 1 - $db_type");
+  ok($o2->last_modified eq $o->last_modified, "save() verify 2 - $db_type");
+  is($o2->start->ymd, '2001-05-24', "save() verify 3 (date value) - $db_type");
+
+  my $o3 = MySQLiteObject->new();
+
+  my $db = $o3->db or die $o3->error;
+
+  ok(ref $db && $db->isa('Rose::DB'), "db() - $db_type");
+
+  is($db->dbh, $o3->dbh, "dbh() - $db_type");
+
+  my $o4 = MySQLiteObject->new(id => 999);
+  ok(!$o4->load(speculative => 1), "load() nonexistent - $db_type");
+  ok($o4->not_found, "not_found() 2 - $db_type");
+
+  ok($o->delete, "delete() - $db_type");
+
+  eval { $o->meta->alias_column(nonesuch => 'foo') };
+  ok($@, 'alias_column() nonesuch');
+}
+
 BEGIN
 {
   #
@@ -638,6 +744,76 @@ EOF
     MyInformixObject->meta->alias_column(save => 'save_col');
     MyInformixObject->meta->initialize(preserve_existing => 1);
   }
+
+  #
+  # SQLite
+  #
+
+  eval 
+  {
+    $dbh = Rose::DB->new('sqlite_admin')->retain_dbh()
+      or die Rose::DB->error;
+  };
+
+  if(!$@ && $dbh)
+  {
+    our $HAVE_SQLITE = 1;
+
+    # Drop existing table and create schema, ignoring errors
+    {
+      local $dbh->{'RaiseError'} = 0;
+      local $dbh->{'PrintError'} = 0;
+      $dbh->do('DROP TABLE rose_db_object_test');
+    }
+
+    $dbh->do(<<"EOF");
+CREATE TABLE rose_db_object_test
+(
+  id             INTEGER PRIMARY KEY AUTOINCREMENT,
+  name           VARCHAR(32) NOT NULL,
+  flag           BOOLEAN NOT NULL,
+  flag2          BOOLEAN,
+  status         VARCHAR(32) DEFAULT 'active',
+  bits           VARCHAR(5) DEFAULT '00101' NOT NULL,
+  start          DATE,
+  save           INT,
+  last_modified  TIMESTAMP,
+  date_created   TIMESTAMP
+)
+EOF
+
+    $dbh->disconnect;
+
+    # Create test subclass
+
+    package MySQLiteObject;
+
+    our @ISA = qw(Rose::DB::Object::Std::Cached);
+
+    sub init_db { Rose::DB->new('sqlite') }
+
+    MySQLiteObject->meta->table('rose_db_object_test');
+
+    MySQLiteObject->meta->columns
+    (
+      'name',
+      id       => { primary_key => 1 },
+      flag     => { type => 'boolean', default => 1 },
+      flag2    => { type => 'boolean' },
+      status   => { default => 'active' },
+      start    => { type => 'date', default => '12/24/1980' },
+      save     => { type => 'scalar' },
+      bits     => { type => 'bitfield', bits => 5, default => 101 },
+      last_modified => { type => 'timestamp' },
+      date_created  => { type => 'timestamp' },
+    );
+
+    eval { MySQLiteObject->meta->initialize };
+    Test::More::ok($@, 'meta->initialize() reserved method');
+
+    MySQLiteObject->meta->alias_column(save => 'save_col');
+    MySQLiteObject->meta->initialize(preserve_existing => 1);
+  }
 }
 
 END
@@ -672,6 +848,17 @@ END
   {
     # Informix
     my $dbh = Rose::DB->new('informix_admin')->retain_dbh()
+      or die Rose::DB->error;
+
+    $dbh->do('DROP TABLE rose_db_object_test');
+
+    $dbh->disconnect;
+  }
+
+  if($HAVE_SQLITE)
+  {
+    # SQLite
+    my $dbh = Rose::DB->new('sqlite_admin')->retain_dbh()
       or die Rose::DB->error;
 
     $dbh->do('DROP TABLE rose_db_object_test');
